# docs/openapi.yaml
openapi: 3.1.0
info:
  title: KELE API
  version: 0.3.0
  summary: Baseline contract for the KELE storefront and administration API.
  description: |
    This contract is the transport-level source of truth for version 1.
    Customer-facing version 1 content is Persian (fa-IR) and RTL. Monetary
    amounts are authoritative integer IRR values; display strings may use
    toman.
    Operations marked with an open-question reference are provisional and must
    not be exposed as final customer behavior until the question is resolved.
servers:
  - url: /api/v1
    description: Same-origin API
...
paths:
...
  /discovery/occasions:
    get:
      operationId: listPublishedOccasions
      summary: List published Occasion discovery entries.
      tags: [Editorial]
      security: []
      responses:
        '200':
          description: Published Occasion discovery entries in configured order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryPage'
        '400':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
  /catalog/products:
    get:
      operationId: listCatalogProducts
      tags: [Catalog]
      security: []
      summary: List published product color variants.
      parameters:
        - $ref: '#/components/parameters/PageCursor'
        - $ref: '#/components/parameters/PageSize'
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
            maxLength: 120
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc]
      responses:
        '200':
          description: Published catalog page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCardPage'
        default:
          $ref: '#/components/responses/Problem'
  /catalog/products/{slug}:
    get:
      operationId: getCatalogProduct
      tags: [Catalog]
      security: []
      summary: Get a published product by slug.
      parameters:
        - $ref: '#/components/parameters/Slug'
        - name: color
          in: query
          description: Optional color-variant identifier to select initially.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /catalog/categories:
    get:
      operationId: listCatalogCategories
      tags: [Catalog]
      security: []
      summary: List published catalog categories in display order.
      responses:
        '200':
          description: Published category navigation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryPage'
        '400':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /catalog/categories/{slug}:
    get:
      operationId: getCatalogCategory
      tags: [Catalog]
      security: []
      summary: Get one published category and its published product cards.
      parameters:
        - $ref: '#/components/parameters/Slug'
        - $ref: '#/components/parameters/PageCursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Published category detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDetail'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
  /admin/categories:
    get:
      operationId: listAdminCategories
      tags: [Admin Catalog]
      summary: List categories for catalog administration.
      security:
        - SessionCookie: [catalog:read]
      responses:
        '200':
          description: Administrative category list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminCategory'
        '401':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createAdminCategory
      tags: [Admin Catalog]
      summary: Create a draft or published catalog category.
      security:
        - SessionCookie: [catalog:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCategoryInput'
      responses:
        '201':
          description: Draft category created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCategory'
        '409':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /admin/categories/{categoryId}:
    patch:
      operationId: updateAdminCategory
      tags: [Admin Catalog]
      summary: Update versioned Category or Occasion discovery content.
      security:
        - SessionCookie: [catalog:write]
      parameters:
        - $ref: '#/components/parameters/CategoryId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCategoryInput'
      responses:
        '200':
          description: Category discovery content updated and audited.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCategory'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /admin/media:
    get:
      operationId: listAdminMedia
      tags: [Admin Content]
      summary: List reusable media metadata.
      security:
        - SessionCookie: [catalog:read]
      responses:
        '200':
          description: Reusable media assets.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Media'
        '401':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createAdminMedia
      tags: [Admin Content]
      summary: Register reusable media metadata.
      security:
        - SessionCookie: [catalog:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminMediaInput'
      responses:
        '201':
          description: Reusable media metadata registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
        '422':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /admin/products:
    get:
      operationId: listAdminProducts
      tags: [Admin Catalog]
      security:
        - SessionCookie: [catalog:read]
      parameters:
        - $ref: '#/components/parameters/PageCursor'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PublicationStatus'
        - name: search
          in: query
          schema:
            type: string
            maxLength: 120
      responses:
        '200':
          description: Administrative product page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProductPage'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createAdminProduct
      tags: [Admin Catalog]
      security:
        - SessionCookie: [catalog:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminProductInput'
      responses:
        '201':
          description: Draft product created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProduct'
        default:
          $ref: '#/components/responses/Problem'
  /admin/products/{productId}:
    get:
      operationId: getAdminProduct
      tags: [Admin Catalog]
      security:
        - SessionCookie: [catalog:read]
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Administrative product.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProduct'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      operationId: updateAdminProduct
      tags: [Admin Catalog]
      security:
        - SessionCookie: [catalog:write]
      parameters:
        - $ref: '#/components/parameters/ProductId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminProductInput'
      responses:
        '200':
          description: Product updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProduct'
        '409':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /admin/products/{productId}/publish:
    post:
      operationId: publishAdminProduct
      tags: [Admin Catalog]
      security:
        - SessionCookie: [catalog:publish]
      parameters:
        - $ref: '#/components/parameters/ProductId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Product and eligible child state published.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProduct'
        '422':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /admin/products/{productId}/validation:
    get:
      operationId: validateAdminProduct
      tags: [Admin Catalog]
      summary: Validate a product aggregate for publication.
      security:
        - SessionCookie: [catalog:read]
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Publication validation result with stable business-rule IDs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicationValidation'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /admin/products/{productId}/preview:
    get:
      operationId: previewAdminProduct
      tags: [Admin Catalog]
      summary: Preview a product without exposing it publicly.
      security:
        - SessionCookie: [catalog:read]
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Protected storefront-shaped preview of a draft or published product.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProductPreview'
        '404':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /admin/skus/{skuId}/prices:
    post:
      operationId: createAdminSkuPrice
      tags: [Admin Catalog]
      summary: Append a SKU price and update its current projection.
      security:
        - SessionCookie: [pricing:write]
      parameters:
        - $ref: '#/components/parameters/SkuId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminPriceInput'
      responses:
        '201':
          description: Immutable price fact and updated current projection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Money'
        '422':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /admin/products/{productId}/archive:
    post:
      operationId: archiveAdminProduct
      tags: [Admin Catalog]
      security:
        - SessionCookie: [catalog:publish]
      parameters:
        - $ref: '#/components/parameters/ProductId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Product archived without removing history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProduct'
        default:
          $ref: '#/components/responses/Problem'
...
  /admin/inventory/{skuId}:
    get:
      operationId: getAdminInventory
      tags: [Admin Inventory]
      security:
        - SessionCookie: [inventory:read]
      parameters:
        - $ref: '#/components/parameters/SkuId'
      responses:
        '200':
          description: Current inventory projection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inventory'
        default:
          $ref: '#/components/responses/Problem'
  /admin/inventory/{skuId}/actions:
    post:
      operationId: createAdminInventoryAction
      tags: [Admin Inventory]
      description: The canonical receipt binds the key to the SKU and normalized action, quantity and reason; any mismatch returns 409 `IDEMPOTENCY_KEY_REUSED` before inventory mutation. Super Admin may use every listed action, Inventory Admin may use production/manual correction/damaged goods, and Instagram Admin may use only Instagram sale/return.
      security:
        - SessionCookie: [inventory:write]
      parameters:
        - $ref: '#/components/parameters/SkuId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [action, quantity, reason]
              properties:
                action:
                  type: string
                  enum:
                    - production
                    - instagram_sale
                    - instagram_return
                    - manual_correction
                    - damaged_goods
                quantity:
                  type: integer
                  not:
                    const: 0
                reason:
                  type: string
                  minLength: 3
                  maxLength: 500
      responses:
        '201':
          description: Inventory movement and updated projection.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inventory'
        '409':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: kele_session
      description: Secure server-managed customer or administrator session.
    AnonymousCart:
      type: apiKey
      in: cookie
      name: kele_cart
      description: Signed high-entropy anonymous cart identifier.
  parameters:
    PageCursor:
      name: cursor
      in: query
      schema:
        type: string
        maxLength: 500
    PageSize:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 24
    Slug:
      name: slug
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        maxLength: 160
...
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SkuId:
      name: skuId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        High-entropy command replay key. Sensitive Milestone 6 commands persist
        a global SHA-256 fingerprint over fingerprint version, command type,
        target, optimistic resource version (or null) and normalized payload.
        An exact replay returns the current result without another side effect;
        reuse for any different fingerprint returns 409
        `IDEMPOTENCY_KEY_REUSED` before a business mutation.
      schema:
        type: string
        minLength: 16
        maxLength: 120
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Optimistic resource version.
      schema:
        type: string
        pattern: '^"[1-9][0-9]*"$'
...
  responses:
    Problem:
      description: RFC 9457-style problem response.
      headers:
        X-Correlation-Id:
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    PublicationStatus:
      type: string
      enum: [draft, published, archived]
...
    CategoryPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CategorySummary'
    CategoryDetail:
      allOf:
        - $ref: '#/components/schemas/CategorySummary'
        - type: object
          required: [products, seo]
          properties:
            products:
              $ref: '#/components/schemas/ProductCardPage'
            seo:
              $ref: '#/components/schemas/Seo'
    Media:
      type: object
      additionalProperties: false
      required: [id, url, width, height, alt, focalPoint]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri-reference
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
        alt:
          type: string
        focalPoint:
          type: object
          required: [x, y]
          properties:
            x:
              type: number
              minimum: 0
              maximum: 1
            y:
              type: number
              minimum: 0
              maximum: 1
...
    ProductCardPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductCard'
        page:
          $ref: '#/components/schemas/CursorPage'
    CursorPage:
      type: object
      required: [hasMore]
      properties:
        nextCursor:
          type: [string, 'null']
        hasMore:
          type: boolean
...
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/ProductCard'
        - type: object
          required: [description, variants, categories, seo, updatedAt]
          properties:
            description:
              type: string
            details:
              type: array
              items:
                type: string
            variants:
              type: array
              items:
                type: object
                required: [id, name, gallery, skus]
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  gallery:
                    type: array
                    minItems: 1
                    items:
                      $ref: '#/components/schemas/Media'
                  skus:
                    type: array
                    items:
                      $ref: '#/components/schemas/SkuOption'
            categories:
              type: array
              items:
                $ref: '#/components/schemas/CategorySummary'
            seo:
              $ref: '#/components/schemas/Seo'
            updatedAt:
              type: string
              format: date-time
...
    AdminCategoryInput:
      type: object
      additionalProperties: false
      required: [name, slug, displayOrder]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        slug:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
          maxLength: 160
        description:
          type: [string, 'null']
          maxLength: 2000
        displayOrder:
          type: integer
          minimum: 0
        status:
          $ref: '#/components/schemas/PublicationStatus'
        discoveryKind:
          type: string
          enum: [catalog, occasion]
        editorialTitle:
          type: [string, 'null']
          maxLength: 180
        editorialDescription:
          type: [string, 'null']
          maxLength: 2000
        heroMediaId:
          type: [string, 'null']
          format: uuid
        seoTitle:
          type: [string, 'null']
          maxLength: 180
        seoDescription:
          type: [string, 'null']
          maxLength: 320
    AdminCategory:
      allOf:
        - $ref: '#/components/schemas/AdminCategoryInput'
        - type: object
          required: [id, status, version]
          properties:
            id:
              type: string
              format: uuid
            status:
              $ref: '#/components/schemas/PublicationStatus'
            version:
              type: integer
              minimum: 0
    AdminMediaInput:
      type: object
      additionalProperties: false
      required: [url, width, height, alt, format, group]
      properties:
        url:
          type: string
          format: uri-reference
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
        alt:
          type: string
          minLength: 1
          maxLength: 500
        format:
          type: string
          enum: [jpg, png, webp]
        group:
          type: string
          enum: [product_images, outfit_editorial, homepage, journal, shared_assets]
        focalPoint:
          type: object
          additionalProperties: false
          required: [x, y]
          properties:
            x:
              type: number
              minimum: 0
              maximum: 1
            y:
              type: number
              minimum: 0
              maximum: 1
...
    AdminProductInput:
      type: object
      additionalProperties: false
      required: [name, slug, description, categoryIds, variants]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 180
        slug:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
          maxLength: 160
        description:
          type: string
          maxLength: 10000
        categoryIds:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
        seo:
          $ref: '#/components/schemas/Seo'
        details:
          type: array
          maxItems: 20
          items:
            type: string
            maxLength: 500
        variants:
          type: array
          items:
            $ref: '#/components/schemas/AdminColorVariantInput'
    AdminProduct:
      allOf:
        - $ref: '#/components/schemas/AdminProductInput'
        - type: object
          required: [id, status, version, variants]
          properties:
            id:
              type: string
              format: uuid
            status:
              $ref: '#/components/schemas/PublicationStatus'
            version:
              type: integer
              minimum: 0
            variants:
              type: array
              items:
                $ref: '#/components/schemas/AdminColorVariantInput'
    AdminProductPage:
      type: object
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminProduct'
        page:
          $ref: '#/components/schemas/CursorPage'
...
    AdminPriceInput:
      type: object
      additionalProperties: false
      required: [amountRial, reason]
      properties:
        amountRial:
          type: integer
          format: int64
          minimum: 1
        reason:
          type: string
          minLength: 3
          maxLength: 500
...
    PublicationValidation:
      type: object
      additionalProperties: false
      required: [valid, errors]
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            $ref: '#/components/schemas/PublicationValidationError'
    AdminProductPreview:
      type: object
      additionalProperties: false
      required: [preview, product]
      properties:
        preview:
          type: boolean
          const: true
        product:
          $ref: '#/components/schemas/ProductDetail'
...
security:
  - SessionCookie: []